Rename Gateway to DataStoreGateway
