+++ packages/api/src/routes/me.integration.test.ts
@@ -2,11 +2,11 @@ import { UserGroup } from "@moj-bichard7/common/types/UserGroup"
 import type { FastifyInstance } from "fastify"
 import { OK } from "http-status"
 import build from "../app"
-import FakeGateway from "../services/gateways/fakeGateway"
+import FakeDataStoreGateway from "../services/gateways/dataStoreGateways/fakeDataStoreGateway"
 import { generateJwtForStaticUser } from "../tests/helpers/userHelper"
 
 describe("/me", () => {
-  const gateway = new FakeGateway()
+  const gateway = new FakeDataStoreGateway()
   let app: FastifyInstance
 
   beforeAll(async () => {
