+++ packages/api/src/server/auth/authenticate.ts
@@ -2,10 +2,10 @@ import type { User } from "@moj-bichard7/common/types/User"
 import type { FastifyReply, FastifyRequest } from "fastify"
 import { BAD_GATEWAY, UNAUTHORIZED } from "http-status"
 import handleDisconnectedError from "../../services/db/handleDisconnectedError"
-import type Gateway from "../../services/gateways/interfaces/gateway"
+import type DataStoreGateway from "../../services/gateways/interfaces/dataStoreGateway"
 import jwtVerify from "./jwtVerify"
 
-export default async function (gateway: Gateway, request: FastifyRequest, reply: FastifyReply) {
+export default async function (gateway: DataStoreGateway, request: FastifyRequest, reply: FastifyReply) {
   const token = request.headers["authorization"]
 
   if (!token?.startsWith("Bearer ")) {
