+++ packages/api/src/services/gateways/interfaces/dataStoreGateway.ts
@@ -1,8 +1,8 @@
 import type { User } from "@moj-bichard7/common/types/User"
 
-interface Gateway {
+interface DataStoreGateway {
   fetchUserByUsername: (username: string) => Promise<User>
   canCaseBeResubmitted: (username: string, caseId: number, forceIds: number[]) => Promise<boolean>
 }
 
-export default Gateway
+export default DataStoreGateway
